% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moarcolourz.R
\name{moarcolourz}
\alias{moarcolourz}
\title{moarcolourz: A High-Contrast Colour Palette Generator}
\usage{
moarcolourz(n, spectrum = custom_spectrum)
}
\arguments{
\item{n}{Number of colours to generate.}

\item{spectrum}{Function defining the colour spectrum you want to use (default is \code{custom_spectrum}, which includes rainbow, browns, and greys).}
}
\value{
A vector of \code{n} colours in hex format.
}
\description{
I got real sick and tired of having to make high-contrast colour palettes for data visualization, so this will do it for us.
Real easy, you pick the number of colours you want, moarcolourz generates a spectrum and then scrambles it so similar colours are not together.
}
